Remove LANGUAGE_VERSION from non-coroutine codegen box tests
Most of these tests used this directive as a way to opt in to a new
language feature, and most of those features are already stable for a
long time, so no opt-in is needed. Some other tests used the directive
to opt out from a language feature, replace those by the `LANGUAGE`
directive. A few tests used the directive to test behavior that actually
depended on the API version; use `API_VERSION` directive there instead.